fixing PSR and moving interfaces to qiboml.interfaces
