+++ tests/test_models_interfaces.py
@@ -38,29 +38,29 @@ def random_subset(nqubits, k):
 
 
 def build_linear_layer(frontend, input_dim, output_dim):
-    if frontend.__name__ == "qiboml.models.pytorch":
+    if frontend.__name__ == "qiboml.interfaces.pytorch":
         return frontend.torch.nn.Linear(input_dim, output_dim)
-    elif frontend.__name__ == "qiboml.models.keras":
+    elif frontend.__name__ == "qiboml.interfaces.keras":
         return frontend.keras.layers.Dense(output_dim)
     else:
         raise_error(RuntimeError, f"Unknown frontend {frontend}.")
 
 
 def build_sequential_model(frontend, layers, binary=False):
-    if frontend.__name__ == "qiboml.models.pytorch":
+    if frontend.__name__ == "qiboml.interfaces.pytorch":
         activation = frontend.torch.nn.Threshold(1, 0)
         layers = layers[:1] + [activation] + layers[1:] if binary else layers
         return frontend.torch.nn.Sequential(*layers)
-    elif frontend.__name__ == "qiboml.models.keras":
+    elif frontend.__name__ == "qiboml.interfaces.keras":
         return frontend.keras.Sequential(layers)
     else:
         raise_error(RuntimeError, f"Unknown frontend {frontend}.")
 
 
 def random_tensor(frontend, shape, binary=False):
-    if frontend.__name__ == "qiboml.models.pytorch":
+    if frontend.__name__ == "qiboml.interfaces.pytorch":
         tensor = frontend.torch.randint(0, 2, shape) if binary else torch.randn(shape)
-    elif frontend.__name__ == "qiboml.models.keras":
+    elif frontend.__name__ == "qiboml.interfaces.keras":
         tensor = frontend.tf.random.uniform(shape)
     else:
         raise_error(RuntimeError, f"Unknown frontend {frontend}.")
@@ -69,7 +69,7 @@ def random_tensor(frontend, shape, binary=False):
 
 def train_model(frontend, model, data, target):
     max_epochs = 30
-    if frontend.__name__ == "qiboml.models.pytorch":
+    if frontend.__name__ == "qiboml.interfaces.pytorch":
 
         optimizer = torch.optim.Adam(model.parameters())
         loss_f = torch.nn.MSELoss()
@@ -95,7 +95,7 @@ def train_model(frontend, model, data, target):
 
         return avg_grad / len(data)
 
-    elif frontend.__name__ == "qiboml.models.keras":
+    elif frontend.__name__ == "qiboml.interfaces.keras":
 
         optimizer = frontend.keras.optimizers.Adam()
         loss_f = frontend.keras.losses.MeanSquaredError()
@@ -112,7 +112,7 @@ def eval_model(frontend, model, data, target=None):
     loss = None
     outputs = []
 
-    if frontend.__name__ == "qiboml.models.pytorch":
+    if frontend.__name__ == "qiboml.interfaces.pytorch":
         loss_f = torch.nn.MSELoss()
         with torch.no_grad():
             for x in data:
@@ -120,7 +120,7 @@ def eval_model(frontend, model, data, target=None):
             shape = model(data[0]).shape
         outputs = frontend.torch.vstack(outputs).reshape((data.shape[0],) + shape)
 
-    elif frontend.__name__ == "qiboml.models.keras":
+    elif frontend.__name__ == "qiboml.interfaces.keras":
         loss_f = frontend.keras.losses.MeanSquaredError(
             reduction="sum_over_batch_size",
         )
@@ -135,31 +135,31 @@ def eval_model(frontend, model, data, target=None):
 def set_seed(frontend, seed):
     random.seed(seed)
     np.random.seed(seed)
-    if frontend.__name__ == "qiboml.models.pytorch":
+    if frontend.__name__ == "qiboml.interfaces.pytorch":
         frontend.torch.manual_seed(seed)
 
 
 def random_parameters(frontend, model):
-    if frontend.__name__ == "qiboml.models.pytorch":
+    if frontend.__name__ == "qiboml.interfaces.pytorch":
         new_params = {}
         for k, v in model.state_dict().items():
             new_params.update({k: v + frontend.torch.randn(v.shape) / 2})
-    elif frontend.__name__ == "qiboml.models.keras":
+    elif frontend.__name__ == "qiboml.interfaces.keras":
         new_params = [frontend.tf.random.uniform(model.get_weights()[0].shape)]
     return new_params
 
 
 def get_parameters(frontend, model):
-    if frontend.__name__ == "qiboml.models.pytorch":
+    if frontend.__name__ == "qiboml.interfaces.pytorch":
         return {k: v.clone() for k, v in model.state_dict().items()}
-    elif frontend.__name__ == "qiboml.models.keras":
+    elif frontend.__name__ == "qiboml.interfaces.keras":
         return model.get_weights()
 
 
 def set_parameters(frontend, model, params):
-    if frontend.__name__ == "qiboml.models.pytorch":
+    if frontend.__name__ == "qiboml.interfaces.pytorch":
         model.load_state_dict(params)
-    elif frontend.__name__ == "qiboml.models.keras":
+    elif frontend.__name__ == "qiboml.interfaces.keras":
         model.set_weights(params)
 
 
@@ -182,7 +182,7 @@ def backprop_test(frontend, model, data, target):
 
 @pytest.mark.parametrize("layer,seed", zip(ENCODING_LAYERS, [1, 4]))
 def test_encoding(backend, frontend, layer, seed):
-    if frontend.__name__ == "qiboml.models.keras":
+    if frontend.__name__ == "qiboml.interfaces.keras":
         pytest.skip("keras interface not ready.")
     if backend.name not in ("pytorch", "jax"):
         pytest.skip("Non pytorch/jax differentiation is not working yet.")
@@ -222,7 +222,7 @@ def test_encoding(backend, frontend, layer, seed):
 @pytest.mark.parametrize("layer,seed", zip(DECODING_LAYERS, [1, 2, 1, 1]))
 @pytest.mark.parametrize("analytic", [True, False])
 def test_decoding(backend, frontend, layer, seed, analytic):
-    if frontend.__name__ == "qiboml.models.keras":
+    if frontend.__name__ == "qiboml.interfaces.keras":
         pytest.skip("keras interface not ready.")
     if backend.name not in ("pytorch", "jax"):
         pytest.skip("Non pytorch/jax differentiation is not working yet.")
